+++ packages/orq-rc/src/models/operations/listagents.ts
@@ -1070,7 +1070,7 @@ export const ListAgentsKnowledgeBases$inboundSchema: z.ZodType<
   z.ZodTypeDef,
   unknown
 > = z.object({
-  id: z.string().default("01K7GJX5TAXDK94ZR5F3NS7R4Y"),
+  id: z.string().default("01K7H8V1G0Q0702BX3SPYC23E4"),
   knowledge_id: z.string(),
   configuration: z.union([
     z.lazy(() =>
@@ -1101,7 +1101,7 @@ export const ListAgentsKnowledgeBases$outboundSchema: z.ZodType<
   z.ZodTypeDef,
   ListAgentsKnowledgeBases
 > = z.object({
-  id: z.string().default("01K7GJX5TAXDK94ZR5F3NS7R4Y"),
+  id: z.string().default("01K7H8V1G0Q0702BX3SPYC23E4"),
   knowledgeId: z.string(),
   configuration: z.union([
     z.lazy(() =>